fix: move tests:e2e to general folder
