+++ cucumber.js
@@ -1,0 +1,12 @@
+// cucumber.js
+let common = [
+    'src/__tests-e2e__/features/**/*.feature', // Specify our feature files
+    '--require-module ts-node/register',       // Load TypeScript module
+    '--require step-definitions/**/*.ts',      // Load step definitions
+    '--format progress-bar',                   // Load custom formatter
+    '--format node_modules/cucumber-pretty'    // Load custom formatter
+  ].join(' ');
+  
+  module.exports = {
+    default: common
+  };
