+++ package.json
@@ -14,6 +14,7 @@
     "build": "npm run clean && tsc",
     "clean": "rimraf dist",
     "test": "jest --verbose src",
+    "test:e2e": "tsc && ./node_modules/.bin/cucumber-js -p default",
     "coverage": "jest --collect-coverage --verbose src",
     "release": "standard-version",
     "secrets": "git secrets --scan"
@@ -26,9 +27,15 @@
   "devDependencies": {
     "@commitlint/cli": "^9.1.2",
     "@commitlint/config-conventional": "^9.1.1",
+    "@types/chai": "^4.2.12",
+    "@types/cucumber": "^6.0.1",
     "@types/jest": "^26.0.7",
     "@typescript-eslint/eslint-plugin": "^3.7.0",
     "@typescript-eslint/parser": "^3.7.0",
+    "chai": "^4.2.0",
+    "cucumber": "^6.0.5",
+    "cucumber-pretty": "^6.0.0",
+    "cucumber-tsflow": "^3.4.0",
     "eslint": "^7.5.0",
     "eslint-config-airbnb-base": "^14.2.0",
     "eslint-config-prettier": "^6.11.0",
@@ -43,6 +50,8 @@
     "rimraf": "^3.0.2",
     "standard-version": "^8.0.2",
     "ts-jest": "^26.1.3",
+    "ts-node": "^8.10.2",
+    "tsconfig-paths": "^3.9.0",
     "typescript": "^3.9.7"
   },
   "dependencies": {
