+++ src/__tests-e2e__/world/calculatorWorld.ts
@@ -1,4 +1,4 @@
-import { Calculator } from "../../src/calculator";
+import { Calculator } from "../support/calculator";
 import { World } from "cucumber";
  
 declare module "cucumber" {
