--- tests/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "tests",
-  "version": "1.0.0",
-  "description": "integration tests",
-  "main": "index.js",
-  "scripts": {
-    "test": "tsc && cucumber-js features/"
-  },
-  "author": "Lisandro Gurman",
-  "license": "ISC",
-  "devDependencies": {
-    "@types/chai": "^4.2.12",
-    "@types/cucumber": "^6.0.1"
-  },
-  "dependencies": {
-    "axios": "^0.16.2"
-  },
-  "build": "tsc"
-}
