+++ tsconfig.json
@@ -11,7 +11,8 @@
     "baseUrl": ".",
     "paths": {
       "*": ["node_modules/*", "src/types/*"]
-    }
+    },
+    "experimentalDecorators": true
   },
   "include": ["src/**/*"]
 }
